% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.01.model.hand.R
\name{aic.summary}
\alias{aic.summary}
\title{Sumamrize aic in each model}
\usage{
aic.summary(result.obj)
}
\arguments{
\item{result.obj}{SDAM object}
}
\description{
Return ...
}
\examples{

}
\author{
Tomoki Yasuhara
}
