% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.04.area.D.R
\name{area.D}
\alias{area.D}
\title{Calculate densities in each area}
\usage{
area.D(leg.D.res)
}
\arguments{
\item{leg.D.res}{result of densities in each leg}
}
\description{
Calculate densities in each area based on leg.D result object
}
\examples{


}
\author{
Hiroaki Murata
}
