% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.03.area.D.Table.R
\name{area.D.table}
\alias{area.D.table}
\title{Create table of area density}
\usage{
area.D.table(best.list)
}
\arguments{
\item{best.list}{}
}
\description{
Summarize densities of area in each debris type into the table
}
\examples{

}
\author{
Tomoki Yasuhara
}
