% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08.00.Data.handling.R
\name{Data.check}
\alias{Data.check}
\title{Check data type}
\usage{
Data.check(Sight.Data, numeric_covariate = c("occo"),
  factor_covariate = c("weather", "size"))
}
\arguments{
\item{Sight.Data}{Sighting data of debris}

\item{numeric_covariate}{character vector of covariate to be used as numeric}

\item{factor_covariate}{character vector of covariate to be used as factor}

\item{Effort.Data}{Effort data of survey}
}
\description{
Check and redefine data type for covariate
}
\examples{

}
\author{
Tomoki Yasuhara
}
