% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08.00.Data.handling.R
\name{Effort.hand}
\alias{Effort.hand}
\title{Create effort data for MALIA}
\usage{
Effort.hand(Data.name, Rev.name)
}
\arguments{
\item{Rev.name}{}
}
\description{
The data for malia package is craeated from 'raw' data
}
\examples{

}
\author{
Tomoki Yasuhara
}
