% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.01.TCDM.R
\name{factor.arrange}
\alias{factor.arrange}
\title{Rearrange categorical levels}
\usage{
factor.arrange(Sight.Data, colist)
}
\arguments{
\item{colist}{}
}
\description{
Rearrange categorical data.
}
\examples{

}
\author{
Tomoki Yasuhara
}
