% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.02.grid.D.plot.R
\name{grid.D.plot}
\alias{grid.D.plot}
\title{Plot densities of grids}
\usage{
grid.D.plot(grid.D.res, xl = c(120, 180), yl = c(10, 50),
  Type = NULL, text.size = 6, save = F)
}
\arguments{
\item{grid.D.res}{grid.D object}

\item{xl}{ranges of longitude to be plotted}

\item{yl}{ranges of latitude to be plotted}

\item{Type}{debris type is also plotted in the figure}

\item{text.size}{size of Type}

\item{save}{defoult is set as false, if true, the figure will be saves in current directory}
}
\description{
Plot densities of grids
}
\examples{

}
\author{
Hiroaki Murata
}
