% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.03.grid.D.R
\name{grid.D}
\alias{grid.D}
\title{Calculate densities in each grid}
\usage{
grid.D(leg.D.res, divide = 0.5, ignore = mean(leg.D.res$Leg.Length) *
  0.3)
}
\arguments{
\item{leg.D.res}{result of densities in each leg}

\item{ignore}{}
}
\description{
Calculate densities in each grid based on leg.D result object
}
\examples{


}
\author{
Hiroaki Murata
}
