% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.02.grid.D.Table.R
\name{grid.D.table}
\alias{grid.D.table}
\title{Create table of grid densities}
\usage{
grid.D.table(best.list)
}
\arguments{
\item{best.list}{}
}
\description{
Summarize densities of grids in each debris type into the table
}
\examples{

}
\author{
Tomoki Yasuhara
}
