% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHN.R
\name{HHN}
\alias{HHN}
\title{Half half normal detection function}
\usage{
HHN(x, para, c.p)
}
\arguments{
\item{x}{vector of perpendicular distance}

\item{para}{parameter vector of detection function (sigma,delta)}

\item{c.p}{cut point}
}
\description{
Half half normal detection function
}
\examples{
HHN(10,c(20,10),10)
}
\author{
Hiroaki Murata
}
