% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHR.R
\name{HHR}
\alias{HHR}
\title{Half hazard rate detection function}
\usage{
HHR(x, para, c.p)
}
\arguments{
\item{x}{vector of perpendicular distance}

\item{para}{parameter vector of detection function (sigma,b,delta)}

\item{c.p}{cut point}
}
\description{
Half hazard rate detection function
}
\examples{
HHR(10,c(20,2,10),10)
}
\author{
Hiroaki Murata
}
