% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HN.R
\name{HN}
\alias{HN}
\title{Half normal detection function}
\usage{
HN(x, para, ...)
}
\arguments{
\item{x}{vector of perpendicular distance}

\item{para}{parameter of detection function (sigma)}
}
\description{
Half normal detection function
}
\examples{
HN(0,20)
}
\author{
Hiroaki Murata
}
