% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{HR}
\alias{HR}
\title{Hazard rate detection function}
\usage{
HR(x, para, ...)
}
\arguments{
\item{x}{vector of perpendicular distance}

\item{para}{parameter vector of detection function (sigma,b)}
}
\description{
Hazard rate detection function
}
\examples{
HR(0,c(20,2))
}
\author{
Hiroaki Murata
}
