% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.02.result.combine.R
\name{leg.D.comb}
\alias{leg.D.comb}
\title{Combine multiple survey results}
\usage{
leg.D.comb(vname)
}
\arguments{
\item{vname}{character vector of Voyage name}
}
\description{
Return ...
}
\examples{

}
\author{
Tomoki Yasuhara
}
