% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.01.leg.D.plot.R
\name{leg.D.plot}
\alias{leg.D.plot}
\title{Plot densities of legs}
\usage{
leg.D.plot(leg.D.res, xl = c(120, 180), yl = c(10, 50), Type = NULL,
  myfill = "gray40", text.size = 6, save = F)
}
\arguments{
\item{leg.D.res}{leg.D object}

\item{xl}{ranges of longitude to be plotted}

\item{yl}{ranges of latitude to be plotted}

\item{Type}{debris type is also plotted in the figure}

\item{myfill}{color of continent}

\item{text.size}{size of Type}

\item{save}{defoult is set as false, if true, the figure will be saves in current directory}
}
\description{
Plot densities of legs
}
\examples{

}
\author{
Tomoki Yasuhara
}
