% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.01.leg.D.Table.R
\name{leg.D.table}
\alias{leg.D.table}
\title{Create table of leg densities}
\usage{
leg.D.table(best.list)
}
\arguments{
\item{best.list}{}
}
\description{
Summarize densities of legs in each debris type into the table
}
\examples{

}
\author{
Tomoki Yasuhara
}
