% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.00.combine.R
\name{ls.}
\alias{ls.}
\title{Combine natural and artificial individuals into same category}
\usage{
ls.(leg.table, Voyage.name, Ar = c("FGN", "FGF", "FGO", "EPS", "PBA",
  "PBO", "FP", "PC", "G", "M", "W", "UO"), Nt = c("SW", "DW", "NO"))
}
\arguments{
\item{Ar}{character vector of artificial product}

\item{Nt}{character vector of natural product}
}
\description{
density ...
}
\examples{

}
\author{
Tomoki Yasuhara
}
