% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/77.77.MALIA.R
\name{MALIA}
\alias{MALIA}
\title{Estimate detection function and calculate density}
\usage{
MALIA(Sight.Data, Effort.Data, key, td, cp = 10, colist = NULL,
  area = T, inits = NULL, para.scale = 0.05, rescale = 10^(-3),
  ignore = 0.01)
}
\arguments{
\item{area}{}
}
\description{
Sequence analysis will be conducted
}
\examples{

}
\author{
Hiroaki Murata and Tomoki Yasuhara
}
