% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.01.model.hand.R
\name{model.extract}
\alias{model.extract}
\title{Return survey information}
\usage{
model.extract(aic.mat, result.obj, Order = rep(1, length(aic.mat)))
}
\arguments{
\item{aic.mat}{.debris data}

\item{result.obj}{.effort data}

\item{Order}{Order can designate ranking of model which we have an interest}
}
\description{
Return debris and voyage information as a list type object
}
\examples{

}
\author{
Tomoki Yasuhara
}
