% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08.00.Data.handling.R
\name{Sight.hand}
\alias{Sight.hand}
\title{Create sighting data for MALIA}
\usage{
Sight.hand(Data.name, Rev.name, Breaks = c(0, 5, 10, 15, 20, 25, 30, 35,
  40, 45, 50, 75, 100, 150, 200))
}
\arguments{
\item{Breaks}{}
}
\description{
The data for malia package is craeated from 'raw' data
}
\examples{

}
\author{
Tomoki Yasuhara
}
