% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.00.survey.plot.R
\name{survey.plot}
\alias{survey.plot}
\title{Plot survey track line}
\usage{
survey.plot(Effort.Data, xl = c(120, 180), yl = c(10, 50),
  col = "orange", myfill = "gray40", multi = F)
}
\arguments{
\item{xl}{ranges of longitude to be plotted}

\item{yl}{ranges of latitude to be plotted}

\item{col}{color of trackline}

\item{myfill}{color of continent}

\item{multi}{if True, multiple surveys will be plotted with legend}
}
\description{
Plot survey track line
}
\examples{

}
\author{
Tomoki Yasuhara
}
