% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.01.TCDM.R
\name{TCDM}
\alias{TCDM}
\title{Estimate detection function}
\usage{
TCDM(Voyage.name, type = NULL, key, td, cp = 10, colist = NULL,
  inits = NULL, area = T, para.scale = 0.05, rescale = 10^(-3),
  ignore = 0.01)
}
\arguments{
\item{area}{}
}
\description{
Inner function of SDAM.
}
\examples{

}
\author{
Tomoki Yasuhara
}
