% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.03.result.An.combine.R
\name{vanc}
\alias{vanc}
\title{Combine multiple survey results}
\usage{
vanc(vname, Ar = c("FGN", "FGF", "FGO", "EPS", "PBA", "PBO", "FP", "PC",
  "G", "M", "W", "UO"), Nt = c("SW", "DW", "NO"))
}
\arguments{
\item{vname}{character vector of Voyage name}

\item{Ar}{character vector of artificial product}

\item{Nt}{character vector of natural product}
}
\description{
Artificial
}
\examples{

}
\author{
Tomoki Yasuhara
}
