% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.00.survey.Table.R
\name{survey.inf}
\alias{survey.inf}
\title{Return survey information}
\usage{
survey.inf(Sight.Data, Effort.Data, Artificial = c("FGN", "FGF", "FGO",
  "EPS", "PBA", "PBO", "FP", "PC", "G", "M", "W", "UO"),
  Natural = c("SW", "DW", "NO"))
}
\arguments{
\item{Sight.Data}{.debris data}

\item{Effort.Data}{.effort data}

\item{Artificial}{character vector of artificial product}

\item{Natural}{character vector of natural product}
}
\description{
Return debris and voyage information as a list type object
}
\examples{

}
\author{
Tomoki Yasuhara
}
